fix crash when loading buggy scenarios/saves
